:wrench: Created a new function called WriteStructure for Responses, which allows to write structs directly to packets. :wrench: Changed keep alive IDs' type to long (formerly varint).

+++ server/client.go
@@ -47,7 +47,7 @@ func (reader FullReader) Read(p []byte) (int, error) {
 
 type KeepAliveData struct {
 	Deadline time.Time
-	ID       int32
+	ID       int64
 }
 
 type Connection struct {
